requirements.txt
streamlit
openai
chromadb
tenacity
pdfminer.six
